requirements.txt
flask
google-cloud-storage
opencv-python-headless
requests
numpy
gunicorn
pip install google-cloud-firestore
